[Refactor] 3주차 - Member -> User 클래스 명 수정
