+++ contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/controller/UserController.java
@@ -2,7 +2,7 @@ package com.example.todolist.controller;
 
 import com.example.todolist.dto.ResponseDto;
 import com.example.todolist.dto.userdto.UserRequestDto.UserJoinReqDto;
-import com.example.todolist.service.MemberService;
+import com.example.todolist.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,13 +16,13 @@ import static com.example.todolist.dto.userdto.UserResponseDto.*;
 @Controller
 @RequiredArgsConstructor
 @RequestMapping("/user")
-public class MemberController {
+public class UserController {
 
-    private final MemberService memberService;
+    private final UserService userService;
 
     @PostMapping("/join")
     public ResponseEntity<?> joinUser(@RequestBody UserJoinReqDto userJoinReqDto) {
-        UserJoinRespDto userJoinRespDto = memberService.join(userJoinReqDto);
+        UserJoinRespDto userJoinRespDto = userService.join(userJoinReqDto);
         return new ResponseEntity<>(new ResponseDto<>(1,"회원가입 성공",userJoinRespDto), HttpStatus.CREATED);
     }
 }
