+++ contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/exception/todo/TodoExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.example.todolist.exception;
+package com.example.todolist.exception.todo;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
