+++ contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/exception/todo/TodoNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.todolist.exception;
+package com.example.todolist.exception.todo;
 
 import lombok.Getter;
