+++ contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/repository/UserRepository.java
@@ -3,6 +3,6 @@ package com.example.todolist.repository;
 import com.example.todolist.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 }
