+++ contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/service/UserService.java
@@ -1,10 +1,8 @@
 package com.example.todolist.service;
 
 import com.example.todolist.domain.User;
-import com.example.todolist.dto.userdto.UserRequestDto;
-import com.example.todolist.dto.userdto.UserResponseDto;
 import com.example.todolist.dto.userdto.UserResponseDto.UserJoinRespDto;
-import com.example.todolist.repository.MemberRepository;
+import com.example.todolist.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,12 +11,12 @@ import static com.example.todolist.dto.userdto.UserRequestDto.*;
 
 @Service
 @RequiredArgsConstructor
-public class MemberService {
-    private final MemberRepository memberRepository;
+public class UserService {
+    private final UserRepository userRepository;
 
     @Transactional
     public UserJoinRespDto join(UserJoinReqDto userJoinReqDto) {
-        User user = memberRepository.save(userJoinReqDto.changeEntity(userJoinReqDto));
+        User user = userRepository.save(userJoinReqDto.changeEntity(userJoinReqDto));
         return new UserJoinRespDto(user);
     }
 }
